+++ lib/flutter_graph_widget.dart
@@ -41,11 +41,30 @@ class _FlutterGraphWidgetState extends State<FlutterGraphWidget> {
         widget.algorithm.$size.value = context.size;
       }
 
-      graphCpn.overlays.addEntry('vertex', (_, game) {
-        return widget.options?.vertexPanelBuilder!
-                .call(graphCpn.graph.hoverVertex) ??
-            Container();
-      });
+      addVertexOverlays();
+
+      addEdgeOverlays();
+    });
+  }
+
+  /// set overlays callback for vertex data panel on vertex hover.
+  ///
+  /// 设置浮层面板回调，在鼠标悬停时触发
+  void addVertexOverlays() {
+    graphCpn.overlays.addEntry('vertex', (_, game) {
+      if (graphCpn.graph.hoverVertex == null) return const SizedBox();
+      return widget.options?.vertexPanelBuilder!
+              .call(graphCpn.graph.hoverVertex!) ??
+          const SizedBox();
+    });
+  }
+
+  void addEdgeOverlays() {
+    graphCpn.overlays.addEntry('edge', (_, game) {
+      if (graphCpn.graph.hoverEdge == null) return const SizedBox();
+      return widget.options?.edgePanelBuilder!
+              .call(graphCpn.graph.hoverEdge!) ??
+          const SizedBox();
     });
   }