+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.1+6
+- Data panel embedding. ( For edge )
+
 ## 0.0.1+5
 - Keep children vertexes around the parent vertex.
 - Data panel embedding.